Use sensible defaults when creating texture views
To achieve this we need to store the format and usage flags
when creating the swapchain (for its owned images) and when
creating textures ourselves.

Views now default to the format of the source texture and
try to use either the depth or color aspect depending upon
the flags set on the source texture. If this is not suitable the
caller can alway specify these directly.