Start work on beginning a render pass
We take a similar approach to WebGPU in terms of API and internally
for the Vulkan backend we defer the creation of a suitable render pass
(and next framebuffer) until we begin the render pass.

We cache the render passes on the VulkanDevice indexed by a
VulkanRenderPassKey so that we can look it up on subsequent frames.