More work on beginning a render pass
We now also create a framebuffer on demand if a suitable one
cannot be found on the device. We use the Vulkan 1.0 approach
for framebuffers where we create a framebuffer for each swapchain
image (assuming we are rendering to a swapchain image).

In the future we may wish to investigate using imageless
framebuffers and defer the specification of the images to the
beginning of the renderpass. Need to investigate the availability
of this extension though. It is a kind of mid-point between the
1.0 approach and dynamic rendering.