Don't use placement new in Pool
We were explicitly calling the contained type's dtor and then when
the pool was being destroyed, the std::vector was calling it again
leading to a double delete.

For now just use operator[] to assign entries to an existing location.

If later we find this is a bottleneck we can replace with our own array
and placement new/delete and skip over calling the dtor for already
destroyed objects in the pool dtor.

Change-Id: I9c3a9044ba3beee9ec81b7b68f7bc007b90de287
Reviewed-on: https://codereview.kdab.com/c/kdab/toy-renderer/+/122486
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>