Support setting a bind group without binding a pipeline first
This is a valid use case as bind group bindings persist even when
pipelines are changed (if the pipelines are compatible).

To support this we allow the user to provide a pipeline layout handle
to setBindGroup(). In the case where there is already a pipeline bound
the existing logic of getting the layout from that pipeline is still used.

This allows the user full control over when to bind bind groups. For
example, they can bind a per-frame bind group with camera and other
data that is constant for the whole frame just once up front and then
leave it, whilst changing pipelines.

Change-Id: Ic249cdb88a10133ca07f560d323a39d02a4d47f6
Reviewed-on: https://codereview.kdab.com/c/kdab/toy-renderer/+/122741
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>