Device: allow to query the Adapter from which the device was created
By making the Device hold a pointer to it.
This is a convenience to be able to access the physical device properties.

To ensure the pointer to the Adapter remains valid after device creation and throughout its
liketime, we now make Instance only return ptr to Adapters to ensure that their lifetime
is tied to the instance which should always outlast that of the device.

Adjust unit tests accordingly.

Prevent creating Adapter

Change-Id: Iaa2587e2419583db0eb01ca64a37ca4ce207eba0
Reviewed-on: https://codereview.kdab.com/c/kdab/toy-renderer/+/123039
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>