Begin implementing the non-blocking buffer upload helper
Almost works but we get a warning about destroying a buffer that
may still be in use by the command buffer doing the transfer.

We will need a fence to keep track of when it is safe to release the
staging buffer.

Longer term we could use sub-allocations of a large persistent
staging buffer for such uploads and use the fences to know when
each region can be reused. But one step at a time.

Change-Id: I9851fdc25cc0f233e5d74ba1b70d20dd2610dec8
Reviewed-on: https://codereview.kdab.com/c/kdab/toy-renderer/+/123336
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>