Add more imgui types and create in example layer
Gradually moving towards a design where an ImGuiItem can be created
and rendered to any given render target and can be fed transformed
mouse/keyboard events.

This will eventually allow us to use ImGuiItem as a simple overlay or
as a way to create a texture that can be rendered into 3D world space
as we wish. Coupled with the technique we used in Qt3D to project the
mouse input events we should be able to use ImGui for in-world UIs.

Change-Id: Ie8deb8e29d830b688bfa1efe52d7270b96ee13ed
Reviewed-on: https://codereview.kdab.com/c/kdab/toy-renderer/+/124723
Tested-by: Continuous Integration <build@kdab.com>
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>