Use an env var to find the library assets
For the imgui shaders and the default font the kdgpu_kdgui library needs
to find the asset directory. This is different depending upon your use case.

In development, we now set the KDGPU_KDGUI_ASSET_PATH env var
to point at the assets dir in the build dir. Then when running examples
during development, the KDGpuKDGui::assetPath() function tries to use
the env var if set.

If the build is destined to be installed, the library will hard-wre the
asset path to be in the install dir directory. It can still be overridden
by way of the environment variable if desired (e.g. to try a different
font).

Still do be done: ensure the assets get installed into the correct
location at install time.

Another approach we can also try for small default assets such as these,
is to compile them into the library binary itself.

Change-Id: Ib1b1af47a4d569551696cf513272287cce2b5885
Reviewed-on: https://codereview.kdab.com/c/kdab/toy-renderer/+/124857
Tested-by: Continuous Integration <build@kdab.com>
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>