Look for assets in share folder
By default, look in the ${INSTALL_DIR}/share/assets if it exists
before using the compile time since this might not be defined (if
not using presets), and might be different once app is installed.

TODO: if apps are built as bundles on macOS, this will be have to
be different.
Change-Id: I0e4c55da44ea8109cb0af6d833ff8b520a8a9dac
Reviewed-on: https://codereview.kdab.com/c/kdab/toy-renderer/+/125851
Tested-by: Continuous Integration <build@kdab.com>
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>