VulkanResourceManager::createRenderPassCommandRecorder
This is a function that is called often and the main generator of
temporary allocations according to heaptrack.
The main culprit was the creation of a vector of texture view handles used to
check whether we already have a matching VulkanFrameBuffer.
This was reworked so that instead of holding and comparing vectors we now
compute a unique hash value based on the added attachments.

Change-Id: Idd749647d9d6001dc2ccfbacf4db775c2bacf4c0
Reviewed-on: https://codereview.kdab.com/c/kdab/toy-renderer/+/125925
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>