Compile in resources needed by KDGpuExample
For the ImGui renderer.

We are using the cmrc - cmake resource compiler which
builds a static lib for a resource set that then gets linked in
to the main library.

Everything needed by ImGui is now compiled into the
KDGpuExample library.

The fonts are also left in the assets directory as they may
well be useful for other examples outside of being used
in ImGui.

Change-Id: I7708b2389dadbbe954dde7b6139d4e1c29c03bed
Reviewed-on: https://codereview.kdab.com/c/kdab/toy-renderer/+/126053
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>