VulkanGraphicsApi: fix createQueueFromExistingVkQueue
We can't require a device handle when calling createQueueFromExistingVkQueue if
in order to create the device we also need the queue. Therefore,
createQueueFromExistingVkQueue will create device less Queues (which isn't
problematic unless you call the async upload functions) and
createDeviceFromExistingVkDevice will then set the device on the queue when the
device is created.

Change-Id: Id9e4d5c21ecb771ea5a5587e3c649ac896594759
Reviewed-on: https://codereview.kdab.com/c/kdab/toy-renderer/+/126822
Tested-by: Continuous Integration <build@kdab.com>
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>