Check if the extension is available before using them on Vulkan
If a default requested extension is not found (e.g. wayland) then print
out a warning instead of letting it hard-crash. This applies to user
extensions too. This seems like a net benefit, if the user is not using
a system with Vulkan validation layers then it'll give a clearer message
that something is definitely off.

Change-Id: I59dfaa3f1ed8c7b0f467991ca684b541681faa19
Reviewed-on: https://codereview.kdab.com/c/kdab/toy-renderer/+/127543
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>