Pick a more sensible minImageCount in the examples
Not all GPUs have a minImageCount of 3, which is the default value
in SwapchainOptions. Instead, choose the next best value - which isn't
the best heuristic but at least it won't crash.

This also adds a new function called getSuitableImageCount that other
applications can use if they want a sensible value without having to
calculate it themselves.

Change-Id: Iaddd3182f35a9bb1242be756249f6a15a9886463
Reviewed-on: https://codereview.kdab.com/c/kdab/toy-renderer/+/127781
Tested-by: Continuous Integration <build@kdab.com>
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>