VulkanResourcesManager: use a single common code path for RenderPass creation
We create RenderPasses in two distinct places within the VulkanResourceManager:
- when creating a GraphicsPipeline
- when creating a RenderPassCommandRecorder

We were using two distinct code paths for each that were doing the same thing
but working off different inputs. This patchs make it so that we use a single
render pass creation function that handles different types of inputs.

Change-Id: I29a3bb72ca02744ccb7968e58231114435632546
Reviewed-on: https://codereview.kdab.com/c/kdab/toy-renderer/+/129144
Tested-by: Continuous Integration <build@kdab.com>
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>