Add support for Timestamp Queries
Since timestamps are introduced into the command flow, we can query the results
at anytime but that doesn't mean the results are ready (since commands might not
have been completed).

It is expected users know about this and either wait for queue/device idle or use
fences to know when the results are valid. Additionally, TimestampQueryRecorder
warns and only return available results.

Task-Id: TOY-53
Change-Id: I69ab167ea17c9ceef02562367ccfd781c9c9dab8
Reviewed-on: https://codereview.kdab.com/c/kdab/toy-renderer/+/130952
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>