Ensure we implement the vulkan_memory_allocator
When fetching vulkan-memory-allocator we would create a static lib that would
hold the implementation of VMA (by building a cpp file that defines
VMA_IMPLEMENTATION).

When building KDGpu without fetch content, we would fail to find the
implementation of the VMA function because we were then missing a
cpp file that defined VMA_IMPLEMENTATION.

To work around this problem, we now make fetch content only create an
interface library whose sole purpose is to provide the VMA header.

Then, as part of KDGpu, we have a vk_mem_alloc.cpp file that takes care of
definining VMA_IMPLEMENTATION so that the implementation gets built.

Change-Id: Ieaa452382404df3e29c7236860e57ebee8052932
Reviewed-on: https://codereview.kdab.com/c/kdab/toy-renderer/+/131079
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>