VulkanDevice/ResourceManager: use dedicated vmaAllocator for external memory
To retrieve external handles to memory allocations, we have to allocate with
an allocator that has specified the external handle types to use by specifying
a VkExportMemoryAllocateInfoKHR on pNext.

Since we delegate the task of allocating to the vulkan-memory-allocator we have
two ways of going about this:
- create a dedicated VmaPool and set a VkExportMemoryAllocateInfoKHR set on the
  VmaPoolCreateInfo::pMemoryAllocateNext
- create a VmaAllocator that was created with the external handle types to expose
  set on VmaAllocatorCreateInfo::pTypeExternalMemoryHandlesTypes

Since creating and managing custom pools is a bit tedious, the later approach was
selected. The only caveat is that we can't perform regular allocations (that we don't want
to expose externally) with such an allocator without yielding validation warnings.
For this reason we keep 2 allocators to handle both cases.

Textures and Buffers allocations where adjusted accordingly and VulkanBuffer/VulkanTexture
record the allocator used for their creation.

Change-Id: I460a0314079437f7c9d82620059d0eb0b6771e1a
Reviewed-on: https://codereview.kdab.com/c/kdab/toy-renderer/+/131540
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>
Tested-by: Continuous Integration <build@kdab.com>