Texture/Buffers: provide allocation size and offsets for external handles
Several Buffers/Textures might end up using the same VkDeviceMemory under the hood
but referencing different offsets. In such a case, the Handle for all these objects
would be the same, only the offset would changed.
Since that offset/size information was missing, we had no way of accessing only the
part of external memory associated with a Buffer/Texture.

Change-Id: I54292a6b30803cef5794ef16482f69848077da3c
Reviewed-on: https://codereview.kdab.com/c/kdab/toy-renderer/+/133270
Tested-by: Continuous Integration <build@kdab.com>
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>