Warn about missing Vulkan layers
Unlike missing extensions, a missing layer (commonly the validation
layer when building a debug build) causes a fatal error when trying to
create the instance. Because of that, it's impossible to tell the user
what was the missing layer because all Vulkan tells is
VK_ERROR_LAYER_NOT_PRESENT.

Now we check the available layers, and warn if they're missing. It also
tells you specifically which layer is not present, and continue
execution if it's missing.

Change-Id: I434b5bdc06840999298da4499afb48d5011d9af5
Reviewed-on: https://codereview.kdab.com/c/kdab/toy-renderer/+/134774
Tested-by: Continuous Integration <build@kdab.com>
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>