Start work on a hello_xr example
And a new XrExampleEngineLayer where we can place most of the
OpenXR specifics for now. Later we can refactor the example engine
layers to share code between the combinations of:

* render loop: simple or advanced
* display: window or openxr device

Might want to look at using composition rather than inheritance here
since we have two axes of variation (so far).

Change-Id: Ic2d3be8ca131b36c7759a84ef2a6d0b3d070d339
Reviewed-on: https://codereview.kdab.com/c/kdab/toy-renderer/+/134964
Reviewed-by: Joshua Benjamin Goins <joshua.goins@kdab.com>
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>