Implement main frame loop
Including the xr frame synchronization and acquiring/releasing
swapchain images.

We do not do anything with the swapchain images yet except
acquire/release them and hand them over for compositing along
with a single projection composition layer.

Need to find a nice way to encapsulate the compositor so that
the scene subclasses can control it.

Change-Id: Ia3e844e20efe64757ce81950fda6777bd8cfeea4
Reviewed-on: https://codereview.kdab.com/c/kdab/toy-renderer/+/135066
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>