Add ability to build KDGpu for iOS
Still uses MoltenVK. It's not nicely packaged though so need some
explicit cmake code to create target for known lib and include paths.

KDUtils does not support iOS yet, so we disable everything that relies
on that. This excludes most tests and examples. Also means we don't
need assets.

Added KDGPU_PLATFORM_IOS and KDGPU_PLATFORM_APPLE (which is also
defined on macOS).

Change-Id: Ia2223b0b2edf58778e6e877c414905f382b5c849
Reviewed-on: https://codereview.kdab.com/c/kdab/toy-renderer/+/136382
Tested-by: Continuous Integration <build@kdab.com>
Reviewed-by: Joshua Benjamin Goins <joshua.goins@kdab.com>