Implement instance polling of events
Now that we encounter the situation where the backend needs to
pass information on to the frontend objects, we have had to
introduce a way to allow this communication. We simply set the
pointer to the frontend object onto the backend object and update
it if the frontend object is ever moved.

We also introduce KDBindings::Property to expose the session
state. This allows consumers to either bind to it or connect to
its signals as they wish.

Change-Id: I82062f10e49d4e4c924750cafabfbf17ca67b586
Reviewed-on: https://codereview.kdab.com/c/kdab/toy-renderer/+/136548
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>