Add hello_xr_multiview example
And add multiview support to the XrProjectionLayer class. If multiview
is requested and supported, then each frame the renderView() function
will be called a single time so that the subclass can render all views
in a single go.

If not using multiview, then the renderView() function will get called once
per view per frame just as it did originally.

Change-Id: I7dfcf3efb7926bc4246ec34c4c0776b1477a6b89
Reviewed-on: https://codereview.kdab.com/c/kdab/toy-renderer/+/137944
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>
Tested-by: Continuous Integration <build@kdab.com>