Fix Android builds for devices that support Vulkan 1.3
In order to build the memory allocator for devices compatible with
Android 12 (such as Oculus Quest 2 and 3), we must constrain its API
usage to version 1.1. However, for devices running Android 13 or above,
with that also support Vulkan 1.3, this causes the allocator to crash
with an assert based on the supported API version.

This can be fixed by constraining the version in the device we create.

Change-Id: I7d1cbad28c11f7d53fd5e429f8b71ce19358ce1b
Reviewed-on: https://codereview.kdab.com/c/kdab/toy-renderer/+/138053
Tested-by: Continuous Integration <build@kdab.com>
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>