VulkanRenderPassKey: take view/resolveView format into account
RenderPasses are compabible between one another when they have compatible attachments,
same number of views and samples.
Attachments are compatible if their layout, load/store operations and formats are
compatible.
Up until now the VulkanRenderPassKey was omitting to account for the attachment format.
This would potentially lead to retrieving a RenderPass we assumed was compatible for a set
of attachments which in reality wasn't since attachment formats were differents.

Change-Id: Id73994f9417b686a0818510c534d1196f57eda5a
Reviewed-on: https://codereview.kdab.com/c/kdab/toy-renderer/+/139401
Tested-by: Continuous Integration <build@kdab.com>
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>