compute_oit_transparency: fix starting index into link list
We treat 0 as the end idx in the SSBO linked list.
We computed the first index into the SSBO as
atomicAdd(alphaFragments.nextIdx, 1);
Which would return 0 for the very first fragment. This would therefore lead to
one alpha fragment not being read since being stored at idx 0.
Therefore, we offset the nodeIdx by 1 to ensure we read all alphaFragments.

Change-Id: Ia6640173d894421fc8c610dd82c6e7c565bc6cfc
Reviewed-on: https://codereview.kdab.com/c/kdab/toy-renderer/+/139772
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>