Move glm compile definitions from the dependencies file to KDGpuExample
These are used by the consumer (in this case, us) to enforce certain
features/rules such as radians, the depth range and more. This was
previously done on the glm target itself, and only in the FetchContent
case. This means it fails if there was glm pre-installed on the system,
nor could this existing solution work because CMake will stop you from
modifying targets that aren't yours.

Now it's set on KDGpuExample itself, which fixes the issue.

Change-Id: I8d63c462919d4ee9f88f9a5e38d011b554bfa5fb
Reviewed-on: https://codereview.kdab.com/c/kdab/toy-renderer/+/140003
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>
Tested-by: Continuous Integration <build@kdab.com>