createRenderPassCommandRecorder: provide clearValue for all attachments
If any attachment is set to be cleared, then we need to provide a
VkClearValue for all attachments, indexed by attachment. Values for
attachments that are not cleared will be ignored.

In the depth_texture_lookup example, stencil buffer is set to
clear and depth was set to load. This led to validation warning
since we weren't providing VkClearValue for all attachments.

Change-Id: Ib93b54c122dfa61fd71dab1f413487e52668f15d
Reviewed-on: https://codereview.kdab.com/c/kdab/toy-renderer/+/140859
Tested-by: Continuous Integration <build@kdab.com>
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>