Don't build glm as a library
We can still link to the glm::glm target though rather than
having to switch to the glm-header-only. In the non-library
configuration, the glm target becomes an INTERFACE target
that depends upon glm-header-only. This simplifies our
downstream targets which can now just link to glm::glm
regardless of how glm is built.

We can therefore go back to using the default upstream
version of glm as this works around the bug where it does
not export symbols on Windows.

Change-Id: Id866b43a4eea90af247b27e775082d5c1f4cb99c
Reviewed-on: https://codereview.kdab.com/c/kdab/toy-renderer/+/141660
Tested-by: Continuous Integration <build@kdab.com>
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>