Fix logical device creation on AMD GPUs
The root problem is that the vkCreateDevice call was failing with an
out of host memory error. By a process of elimination this was found
to be caused by including the device group create info into the
pNext chain for the VkDeviceCreateInfo struct. We should only include
the device group create info when we actually have a group of >1
device.

The way we were handling chaining these structs into the pNext
chain was awkward especially in the presence of optional elements.
This has been reworked to use the VkBaseOutStructure and a generic
lambda to manage the pNext chain. This makes it much easier to
work with.

As a drive-by also make sure that the std layout 430 options get
included in the pNext chain again.

Change-Id: I0ec37aa41ae86c3cd2202887dbbc351dc6721fe8
Reviewed-on: https://codereview.kdab.com/c/kdab/toy-renderer/+/141903
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>
Tested-by: Continuous Integration <build@kdab.com>