AccelerationStructure: only create scratchBuffer when needed
We don't need to keep the scratch buffer alive all the time,
it is only needed when building or updating an AccelerationStructure.
Therefore, we only need to allocate it when building/updating an
acceleration structure in VulkanCommandRecorder::buildAccelerationStructures
and can release it as soon as the build/update has been executed.

Change-Id: Ie75d635feeb111f5be133527f9a4203fb2485003
Reviewed-on: https://codereview.kdab.com/c/kdab/toy-renderer/+/142109
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>