TimeStampQueryRecorder: fix the way we handle active query count
We keeps a list of buckets that tell use the starting query and
count for each TimestampQueryRecorder we have. The previous logic was wrong
as we would append a bucket with a starting query that would always be
the sum of the bucket with highest starting query and count. This would fail
to account for the case where we remove TimestampQueryRecorder
(and associated buckets) and hence have available room at lower starting query.

This patch fixes that and updates the associated unit tests.
The patch also fixes sync validation error that resulted from a missing barrier.

Change-Id: I4c74bd0d6e1b34c8b50d247d7e209e25db59fc8b
Reviewed-on: https://codereview.kdab.com/c/kdab/toy-renderer/+/142213
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>
Tested-by: Continuous Integration <build@kdab.com>