RenderPassCommandRecorderOption: make it possible to specify FrameBuffer array layers
So far we were deducing the array layers for the framebuffer from the first
ColorAttachment's Texture.
In the case of stereo rendering where we have 2 arrayLayers on the swapchain
images, this forced us to have a VkFramebuffer with 2 layers. Yet, we might
still want to render to only one of the layers (e.g draw 2D UI only on one
layer) and have a VkFramebuffer with a single layer. This is where the
additional field in the options can be handy.

Change-Id: I4de08c2492f7259c74c3786b8bcd10c78a23185d
Reviewed-on: https://codereview.kdab.com/c/kdab/toy-renderer/+/143440
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>
Tested-by: Continuous Integration <build@kdab.com>