Fix acceleration structure creation on AMD
VkAccelerationStructureBuildSizesInfoKHR has 3 fields:
- the size for build or update operations
- the size for build operations
- the size for update operations

On AMD, only the first field is set, the others are set to 0.
Since we were using these last 2 fields, buffer creation with a size of 0
would fail.
We know check if any of these 2 sizes are 0, and if so set them to the first
fields. This fixes AccelerationStructure/RayTracing on AMD hardware.

Change-Id: I80845e058b9fc81f2d6be237990b5cbcb59b9421
Reviewed-on: https://codereview.kdab.com/c/kdab/kdgpu/+/147839
Tested-by: Continuous Integration <build@kdab.com>
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>