Passthrough for KDXr
Create a wrapper for XrPassthroughLayerFB in kdxr, and a wrapper for
XrCompositionLayerPassthroughFB in KDGpuExample. The singleton
XrPassthroughFB object is kept inside the resource manager, created and
destroyed as required by passthrough layers.
Compose passthrough layers with the other compositor layers at the end
of a frame in the opennxr session.

The video processing and geometry mapping functionality of
XrPassthroughLayerFB are not yet exposed in our interface.

Add a passthrough layer, and a control to stop and start playback to
the hello_xr example.

Change-Id: I87643c1deaeb1167e0995ef967deff1e55751b0e
Reviewed-on: https://codereview.kdab.com/c/kdab/kdgpu/+/150949
Tested-by: Continuous Integration <build@kdab.com>
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>