Add support for HostImageCopy
Allows to not have to create intermediate staging buffers to copy data to or
from images. Also enables to perform layout transitions directly on the host.

In all cases, those operation behave as coherent meaning that:
- No need for barrier for writes to become visible to device.
- The application must guarantee that any previously submitted command
  that reads from or writes to the texture has completed before
  issuing a layout transitions or issuing copies.

Core in Vulkan 1.4 but available through VK_EXT_host_image_copy for earlier
versions.

Change-Id: I15072e8ba51aff1b2d2e23188a6d2bfbd6f2583a
Reviewed-on: https://codereview.kdab.com/c/kdab/kdgpu/+/152024
Tested-by: Continuous Integration <build@kdab.com>
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>