Cleanup use of features/functions coming from extensions
We were loosely assuming that some extensions were available depending
on the plateform we were compiling on. However this missed the fact
that the extensions are defined by the Vulkan SDK we build against.

Therefore, use of structs/functions depending on a specific extensions
need to be protected by ifdef block that check the extension is indeed
defined.

For some windows extension defined in vulkan_32.h, we can't check for the defines
in KDGpu header files. That is because that would imply including windows.h which we
have to include with NOMINMAX and WIN32_LEAN_AND_MEAN. If included with these
defines, this could potentially cause a conflict for people using KDGPU headers
and that include windows without the defines. Therefore, we have to keep relying
for the KDGPU_PLATFORM_WIN32 in some cases.

Note: we assume Vulkan SDK version >= 1.2
Change-Id: Iebd8372f31305aaff3775fee832c42f24544a897
Reviewed-on: https://codereview.kdab.com/c/kdab/kdgpu/+/152051
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>
Tested-by: Continuous Integration <build@kdab.com>