Improve XR ImGui layers
Allow access to overlay object, allow default UI to be removed,
allow background colour to be changed, provide an intersection test
for determining mouse position, and add a utility function to allow
drawing a basic mouse pointer. Add a layer background colour changer to
the default UI. Fix input events to be delivered to the correct
context.

(Quad layers only for now)

Task-Id: KDG-75
Change-Id: Idc6c2d3973d5396f2a2921d1cfdca2da5653658c
Reviewed-on: https://codereview.kdab.com/c/kdab/kdgpu/+/152391
Tested-by: Continuous Integration <build@kdab.com>
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>