Update KDGpu examples for KDUtils filesystem changes
This is required in order for the Android examples to work properly.
Since shader and texture asset files require to be of the appropriate
storage type, use the standard system asset directory as the default
location for them.

Have the examples use the shared KDGpuExample assetDir folder instead
of each defining an identical assetPath function.

Change-Id: Iba275b88b91bb51c99debbfbf21cc53959a4d4a5
Reviewed-on: https://codereview.kdab.com/c/kdab/kdgpu/+/152924
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>
Tested-by: Continuous Integration <build@kdab.com>