BindGroup/BindGroupLayout: add support for immutableSamplers
When creating a BindGroupLayout of type Sampler or CombinedImageSampler,
we can set the samplers to use directly on the layout as immutable
samplers.

Later, when creating BindGroup with that layout if type is
CombinedImageSampler, we don't need to set a sampler on it.

If we have a BindGroupLayout of type Sampler with an immutable sampler,
the user should not create a BindGroup of type Sampler with a sampler
on it as this will results in errors.

Change-Id: Ic65feeb5b1b49ff5a3092f43c1014c61fb71f8d5
Reviewed-on: https://codereview.kdab.com/c/kdab/kdgpu/+/152969
Tested-by: Continuous Integration <build@kdab.com>
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>