Add operator < for Flags
Semantically it doesn't make a lot of sense to compare bitwise flags
in the general case (depending upon how the flags are defined).

However, it can be very useful to be able to use a flags value as the
key in an ordered associative container such as a std::map.

Specifically, in my case it was for coalescing shader paramters exposed
as push constants into a single push constant range per distinct shader
stage flags value.

Change-Id: I1903bee6a9d3d3309469ebaadae91701f983400f
Reviewed-on: https://codereview.kdab.com/c/kdab/kdgpu/+/154437
Reviewed-by: Alistair Baxter <alistair.baxter@kdab.com>