Fix Device Initialization when using RenderDoc
Recent RenderDoc version (1.20+) were failing when KDGpu was initializing
the Vulkan Device.

This is due to enabling the Vulkan Raytracing Pipeline which is apparently
not supported yet by Renderdoc. We entered a case where the feature was
seen as being available but the matching extension was not supported
at runtime. This is likely a mismatching behavior due to the way the
RenderDoc Vulkan Loader works. Eventually, initialization would fail.
To solve this, we now check that the feature and the matching extension
are available at runtime.

Change-Id: If44e64871e8699d8612a43fd30969a584aa9139d