Add feature to exclude some edges from wireframe
Indicated by setting the vertex attribute to 1.0 to exclude the edge
opposite the vertex and 0.0 to include the edge.

The common case works. Have made an attempt to also get the
robust cases working too by passing through the edge exclude
flags to the fragment shader and including this in the edgeDistance
calculations.

Change-Id: I3b4bcdca3a545dd29680f1721e27561f6677b8cd
Reviewed-on: https://codereview.kdab.com/c/kdab/kdgpu/+/156684
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>