Destroy Framebuffers when attachment TextureViews are destroyed
When destroying a TextureView, destroy any Framebuffer that references
the TextureView as an attachment. We can conveniently use the AttachmentKey
of the FramebufferKey for that.

Note that we can also have Framebuffer that have no attachments (imageless).
Those will only get destroyed when the Device is destroyed. Given this is
a rather unusual case.

Task-Id: KDG-90
Change-Id: I046dda5171daff44246f0a2ead0a78e6c8d43032
Reviewed-on: https://codereview.kdab.com/c/kdab/kdgpu/+/157140
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>